+++ R/data.R
@@ -13,3 +13,11 @@
 #' @format A list of lists
 #' @source See included file "inst/extdata/Flora Quadrat 0.2.xml"
 "fq_raw"
+
+#' Submission data for a subgroup of an ODK Central form.
+#'
+#' The OData response for the subgroup of an ODK Central form.
+#'
+#' @format A list of lists
+#' @source See included file "inst/extdata/Flora Quadrat 0.2.xml"
+"fq_tae"