+++ NEWS.md
@@ -1,12 +1,13 @@
 # ruODK 0.1.0
 
-* Parses metadata, submissions, and handles attachments.
+* Parses metadata, submissions, 
+  and handles attachments (retaining already downloaded attachments).
 * Includes example forms as both `.xml` and `.odbbuild` in `inst/extdata`.
 * Includes example data as retrieved from ODK Central.
 * Includes vignette demonstrating tidying of retrieved data.
+* Handles repeating groups.
 
 Roadmap:
 
-* Handle repeating groups.
 * Handle paginated submissions.
 * Retain already downloaded submissions.